Rename navbar and page-loader files and its functions correctly
