+++ src/index.js
@@ -1,3 +1,3 @@
-import start from './navbar';
+import start from './page-loader';
 
 document.body.appendChild(start());
